Fix setting response code on FastCGI installations
Status code on fastcgi deployments must be set by 'Status' not 'HTTP/' header. The
latter will always be dropped and will block setting 'Status'. It's also not possible
to detect reliably from source what type of deployment is used so the only way to
be sure proper status code is set is to use header's() 3rd param.